requirements.txt
flask>=3.0
gunicorn>=21.0
torch>=2.0
ultralytics>=8.0
flask-sqlalchemy>=3.1
pillow>=10.0
opencv-python-headless>=4.8
pandas
seaborn
tqdm
